backend/api_lambda/code_and_requirements/requirements.txt
Flask==3.0.3
aws-wsgi
psycopg2-binary==2.9.3